- Added so it is possible to scale the map. When going with a width/height less than 600 pixels of the map the preset buttons are removed
First start the new version and the new "keys" will be created in the rotator_settings.ini file. These can now be edited to be redirected to another map and also it is possible to set the width and height of the map. The default is maps/map.png and 600, 600 pixels